Added support for map to same type without need for method template type. Added each method.
The first means that before even for simple transform of data, you needed to set again the type of data,
now that is the default so that if you just transform data (same return type as input), there
is no need to set it up.

The second allows to simply apply a function on all data for its side effects, returning the same
list to continue chaining operations if needed.